Add support for latest versions of Orchard
Since the last time we worked on Orchard there have been 3 new releases
of Orchard (`1.9.3`, `1.10` & `1.10.1`). By default the development project
will use the latest version.

Unfortunately the latest version of Orchard has changed how the third party
dependencies are managed. In the latest version Orchard no longer contains
the third party references in the source code, instead using NuGET to
restore them. In order for the grunt `build` task to work the
`ClickToBuild.cmd` needs to be run in order to run the NuGET restore
command.