No longer require sensitive server details to be stored within the source code
Instead of having sensitive server information kept within source code via
the `deployments` directory, this information is now passed via command
arguments when running `grunt deployment`. The advantage of this is the
source code could be hosted within an open source environment and it is
easier to quickly deploy to alternative environments.